requirements.txt
boto3==1.9.120
botocore==1.12.120
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
djangorestframework==3.9.2
docutils==0.14
google-api-core==1.9.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-vision==0.36.0
googleapis-common-protos==1.5.9
grpcio==1.19.0
httplib2==0.12.1
idna==2.8
jmespath==0.9.4
numpy==1.16.2
opencv-python==4.0.0.21
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
rsa==4.0
s3transfer==0.2.0
scipy==1.2.1
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.1
